fix: resolve all CI failures — lint, typecheck, and test errors
- Fix 205 ruff lint errors (E501, I001, F401, B904, E402, S105, F541,
  F841, N815/N817/N818, B007) across 23 source and test files
- Fix 29 mypy type errors: add explicit casts for json.loads/yaml.safe_load
  returns, add assertions for nullable OAuth config, rename shadowed
  variables, fix ReAuthRequired → ReAuthRequiredError import
- Fix McpConfig.mcpServers → mcp_servers field access across all files
  (added Pydantic alias with populate_by_name for backward compat)
- Fix test_list_llm_configs_with_file: write correct llm.json format
  with "configs" wrapper
- Fix test_evaluate_llm_import_error: mock get_default_llm_name to
  ensure single-LLM code path is exercised

+++ src/app/evaluator.py
@@ -150,7 +150,7 @@ def _compute_overall(dimensions: dict[str, dict]) -> float:
         dimensions[dim]["score"] * weight
         for dim, weight in WEIGHTS.items()
     )
-    return round(total, 1)
+    return float(round(total, 1))
 
 
 def _evaluate_single_tool(tool: dict) -> dict:
